Password module refactored
Added configurable type definitions.
AccountsPassword properly injected in Password module.
AccountsServer is now global to be accessible from other
module.
Context builder removed as it's no longer required by
graphql-modules.
