+++ packages/graphql-api/src/index.ts
@@ -1,3 +1,3 @@
-// export * from './modules';
-// export * from './utils';
-// export * from './models';
+export * from './modules';
+export * from './utils';
+export * from './models';
