+++ packages/graphql-api/src/modules/core/schema/mutation.ts
@@ -1,8 +1,8 @@
 
 import { gql } from 'graphql-modules';
-import { AccountsModuleConfig } from '..';
+import { CoreModuleConfig } from '..';
 
-export default (config: AccountsModuleConfig) => gql(`
+export default (config: CoreModuleConfig) => gql(`
   ${config.extendTypeDefs ? 'extend' : ''} type ${config.rootMutationName || 'Mutation'} {
     impersonate(accessToken: String!, impersonated: ImpersonationUserIdentityInput!): ImpersonateReturn
     refreshTokens(accessToken: String!, refreshToken: String!): LoginResult
