+++ packages/graphql-api/src/modules/password/schema/query.ts
@@ -1,0 +1,8 @@
+import { gql } from "graphql-modules";
+import {CoreModuleConfig} from "../../core";
+
+export default (config: CoreModuleConfig) => gql(`
+  extend type ${config.rootQueryName || 'Query'} {
+    twoFactorSecret: TwoFactorSecretKey
+  }
+`);
