+++ packages/graphql-api/src/schema.ts
@@ -3,8 +3,7 @@
  * Do not import directly in your application.
  */
 import { createApplication } from 'graphql-modules';
-import { coreModule } from './modules/core';
-import { passwordModule } from './modules/password';
+import { CoreModule, CoreModuleConfig, PasswordModule} from './modules';
 
 const accountsServer = {
   getServices: () => ({
@@ -12,12 +11,15 @@ const accountsServer = {
   }),
 }
 
+const config = <CoreModuleConfig>{
+  accountsServer: accountsServer as any
+};
+
 const application = createApplication({
   modules: [
-    coreModule({
-      accountsServer
-    }),
-    passwordModule],
+    CoreModule(config),
+    PasswordModule(config)
+  ],
 });
 
 export default application.schema;
