+++ packages/graphql-api/src/utils/index.ts
@@ -1,3 +1,2 @@
 export * from './authenticated-directive';
 export * from './authenticated-resolver';
-export * from './context-builder';
